requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0

# Distributed computing frameworks
ray[default]>=2.5.0
dask[complete]>=2023.5.0
dask[distributed]>=2023.5.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# System monitoring and utilities
psutil>=5.9.0

# Configuration and data formats
PyYAML>=6.0
json5>=0.9.0

# Cloud and deployment
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.24.0

# Optional: Additional ML utilities
scikit-learn>=1.3.0
tqdm>=4.65.0